# Conflicts:
#	introspection-engine/introspection-engine-tests/tests/tables/mod.rs

+++ migration-engine/connectors/sql-migration-connector/src/flavour/mysql.rs
@@ -3,6 +3,7 @@ use crate::{
     connect,
     connection_wrapper::Connection,
     error::{quaint_error_to_connector_error, SystemDatabase},
+    SqlMigrationConnector,
 };
 use datamodel::{walkers::walk_scalar_fields, Datamodel};
 use enumflags2::BitFlags;
@@ -43,6 +44,51 @@ impl MysqlFlavour {
         BitFlags::<Circumstances>::from_bits(self.circumstances.load(Ordering::Relaxed))
             .unwrap_or_default()
             .contains(Circumstances::LowerCasesTableNames)
+    }
+
+    async fn shadow_database_connection(
+        &self,
+        main_connection: &Connection,
+        connector: &SqlMigrationConnector,
+        temporary_database_name: Option<&str>,
+    ) -> ConnectorResult<Connection> {
+        if let Some(shadow_database_connection_string) = &connector.shadow_database_connection_string {
+            let conn = crate::connect(shadow_database_connection_string).await?;
+            let shadow_conninfo = conn.connection_info();
+            let main_conninfo = main_connection.connection_info();
+
+            if shadow_conninfo.host() == main_conninfo.host() && shadow_conninfo.dbname() == main_conninfo.dbname() {
+                return Err(ConnectorError::generic(anyhow::anyhow!("The shadow database you configured appears to be the same as as the main database. Please specify another shadow database.")));
+            }
+
+            tracing::info!(
+                "Connecting to user-provided shadow database at {}",
+                shadow_database_connection_string
+            );
+
+            if self.reset(&conn).await.is_err() {
+                connector.best_effort_reset(&conn).await?;
+            }
+
+            return Ok(conn);
+        }
+
+        let database_name = temporary_database_name.unwrap();
+        let create_database = format!("CREATE DATABASE `{}`", database_name);
+
+        main_connection
+            .raw_cmd(&create_database)
+            .await
+            .map_err(ConnectorError::from)
+            .map_err(|err| err.into_shadow_db_creation_error())?;
+
+        let mut temporary_database_url = self.url.url().clone();
+        temporary_database_url.set_path(&format!("/{}", database_name));
+        let temporary_database_url = temporary_database_url.to_string();
+
+        tracing::debug!("Connecting to temporary database at {:?}", temporary_database_url);
+
+        Ok(crate::connect(&temporary_database_url).await?)
     }
 }
 
@@ -226,28 +272,18 @@ impl SqlFlavour for MysqlFlavour {
         Ok(())
     }
 
-    #[tracing::instrument(skip(self, migrations, connection))]
+    #[tracing::instrument(skip(self, migrations, connection, connector))]
     async fn sql_schema_from_migration_history(
         &self,
         migrations: &[MigrationDirectory],
         connection: &Connection,
+        connector: &SqlMigrationConnector,
     ) -> ConnectorResult<SqlSchema> {
-        let database_name = format!("prisma_shadow_db{}", uuid::Uuid::new_v4());
-        let create_database = format!("CREATE DATABASE `{}`", database_name);
+        let temporary_database_name = connector.temporary_database_name();
 
-        connection
-            .raw_cmd(&create_database)
-            .await
-            .map_err(ConnectorError::from)
-            .map_err(|err| err.into_shadow_db_creation_error())?;
-
-        let mut temporary_database_url = self.url.url().clone();
-        temporary_database_url.set_path(&format!("/{}", database_name));
-        let temporary_database_url = temporary_database_url.to_string();
-
-        tracing::debug!("Connecting to temporary database at {:?}", temporary_database_url);
-
-        let temp_database = crate::connect(&temporary_database_url).await?;
+        let temp_database = self
+            .shadow_database_connection(connection, connector, temporary_database_name.as_deref())
+            .await?;
 
         // We go through the whole process without early return, then clean up
         // the temporary database, and only then return the result. This avoids
@@ -275,8 +311,10 @@ impl SqlFlavour for MysqlFlavour {
         })()
         .await;
 
-        let drop_database = format!("DROP DATABASE IF EXISTS `{}`", database_name);
-        connection.raw_cmd(&drop_database).await?;
+        if let Some(database_name) = temporary_database_name {
+            let drop_database = format!("DROP DATABASE IF EXISTS `{}`", database_name);
+            connection.raw_cmd(&drop_database).await?;
+        }
 
         sql_schema_result
     }
